small changes 2
